--- controller/pyproject.toml
@@ -1,15 +1,0 @@
-[tool.poetry]
-name = "controller"
-version = "0.1.0"
-description = ""
-authors = ["Charles Ueltschey <chaseuelt@gmail.com>"]
-
-[tool.poetry.dependencies]
-python = "^3.10"
-
-[tool.poetry.dev-dependencies]
-pytest = "^5.2"
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
