--- controller/ue/test.kv
@@ -1,8 +1,0 @@
-<main_display>:
-    BoxLayout:
-        orientation: "vertical"
-        
-        Image:
-              id: imageView
-              source: 'Webimage.png'
-              allow_stretch: True
